move some types from portal to eco-svm-std
