+++ programs/portal/tests/fulfill.rs
@@ -6,9 +6,10 @@ use anchor_lang::{system_program, AnchorSerialize};
 use anchor_spl::associated_token::get_associated_token_address_with_program_id;
 use anchor_spl::token::spl_token;
 use anchor_spl::token_2022::spl_token_2022;
+use eco_svm_std::CHAIN_ID;
 use portal::events::IntentFulfilled;
 use portal::state::FulfillMarker;
-use portal::types::{Call, Calldata, CalldataWithAccounts, Route, CHAIN_ID};
+use portal::types::{Call, Calldata, CalldataWithAccounts, Route};
 use portal::{state, types};
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signer::Signer;
